Commit #5:
Updated the animation component to only update the current animation if the returned subject value (from the movement component) was not equal to "IDLE", and to pause the current animation if it was "IDLE". This stops the player's animation looping infinitely when the player is idle.

Added a battle encounter component and system which is used to initialise a player battle. It works under the same premise as normal collision detection, but it holds a battle delay timer which skips rolling the battle chance to avoid always getting in a battle when colliding with that entity.

Renamed all state Draw(sf::RenderWindow& window) methods to Process() as the window was never used in the method, and process seemed like a more appropriate name.

Removed the BattleSystem and moved the needed methods into the BattleState. The BattleSystem didn't need to be a system as it never ran over a series of entities, instead dealt with specific entities.

Added a second Map which will be used as the demo game map - although still incomplete.

Renamed MapLoader to FileLoader

Added music to game.cpp

Added some more basic functionality to the InventoryComponent

Added InventoryMenuSystem, works similar to the already existing menu system, but this is specific to the needs of the inventory system. Can scroll through the list of items in the inventory and equip them to the player, altering their stats. The "Use" option still isn't implemented and the current number of menu text entities is fixed at 4. Needs more work and some polish.

Added a large number of sprites sourced online.

Added a particle emitter class which is used to create/spawn particles given a series of inputs. Particles are made up of a PositionComponent, VelocityComponent, DelayComponent and SpriteComponent. Added angle and angular velocity to the VelocityComponent to accommodate for particles. The delay component has a clock and a delay time, when the particle system sees that the delay clock time exceeds the delay time, the entity is removed from the entityManager.

Added an isOnScreen method to the SpriteSystem so it will now only draw an entity if it is on the screen, freeing up some resources.